Migrate cloud storage panel into storage settings
